Refactor graphql queries for performance
Previously I was doing a query on the entire set of metadata fields and then modifying the elements and reloading the whole query. Now I am simply doing the big query on application load, and then updating individual fields as needed. This prevents a full tab rerender on every individual update.