Use the "render props" pattern
I spent a little extra time on this and I have managed to successfully use the ["render props" pattern](https://reactjs.org/docs/render-props.html) so that there is a `MetadataContainer` component that handles all the graphql and save/edit/delete UI logic (since that's the same across components). Then that component itself gets passed different `render` functions conditionally based on the data type it's supposed to be rendering. So now each individual data type component is strictly supposed to render data it is passed -- the `MetadataContainer` component is where any magic happens.