Completed DateTime component
Using the `@material-ui/pickers` component since there is not great native UX for date picking, and this way it will be consistent across browsers. Default time zone is GMT -- I determined this by testing what happens on production `check-web` if I don't pick any metadata field and enter a time. Even if GMT is not one of the "allowed" timezones, if a user enters nothing, the default is GMT. So I replicated that behavior here.