Add conditional metadata rendering
Conditional fields are now rendered based on whether they meet their preconditions.

Also added was a "setRerender" function, which is called whenever an individual metadata item is updated by the user. This rerenders the parent list of all metadata items (since the items rendered are dependent on the results of individual items now). This is handled a little differently in check-web, but this method seems like a more "relay modern" way of doing it (aside from re-querying the entire list of items).